Updated Harsh/Idle states, removed states and images Harsh states and idle states are now in Trip object
